+++ src/main/java/com/mercadolibre/planning/model/api/domain/usecase/unitsdistribution/create/UnitsInput.java
@@ -16,7 +16,7 @@ public class UnitsInput {
 
     String area;
 
-    long quantity;
+    Double quantity;
 
     String quantityMetricUnit;
 }
