+++ src/main/java/com/mercadolibre/planning/model/api/web/controller/unitsdistibution/request/Units.java
@@ -24,7 +24,7 @@ public class Units {
     private String area;
 
     @NotBlank
-    private long quantity;
+    private Double quantity;
 
     @NotBlank
     private String quantityMetricUnit;
